split astyle_wrappers into multiple files
