+++ astyle_py/__init__.py
@@ -1,1 +1,8 @@
-__version__ = "0.1.0"
+# SPDX-FileCopyrightText: 2022 Ivan Grokhotkov <ivan@igrr.me>
+# SPDX-License-Identifier: MIT
+from .astyle_wrapper import Astyle
+
+__all__ = [
+    'Astyle',
+]
+__version__ = '0.2.0'
