+++ setup.cfg
@@ -25,7 +25,7 @@ astyle_py = libastyle.wasm
 
 [options.entry_points]
 console_scripts =
-    astyle_py = astyle_py.astyle_wrapper:main
+    astyle_py = astyle_py.__main__:main
 
 [bdist_wheel]
 universal = True
